+++ src/components/cards/Order/OrderHeaderCard.vue
@@ -1,19 +1,32 @@
 <template>
 <div class="content-heading-wrap proudct-heading-wrap">
 		<div class="content-heading-wrap-elem">
-		<div class="content-heading">Заказ № 2435</div>
+		<div class="content-heading">Заказ № {{data.id}}</div>
 		</div>
 		<div class="content-heading-wrap-elem">
 		<div class="content-heading-price"> 
 			<div class="content-heading-price-text">Сумма заказа: </div>
-			<div class="content-heading-price-value">547 300 ₽</div>
+			<div class="content-heading-price-value">{{Number(data.total).toLocaleString()}} ₽</div>
 		</div>
 		</div>
 		<div class="content-heading-info"> 
-		<div class="content-heading-info-elem"> <span class="content-heading-info-text">Колечество едениц: </span><span class="content-heading-info-value">120</span></div>
-		<div class="content-heading-info-elem"> <span class="content-heading-info-text">Общий вес: </span><span class="content-heading-info-value">25 кг.</span></div>
-		<div class="content-heading-info-elem"> <span class="content-heading-info-text">Общий объем: </span><span class="content-heading-info-value">23 м.куб.</span></div>
+		<div class="content-heading-info-elem"> <span class="content-heading-info-text">Колечество едениц: </span><span class="content-heading-info-value">{{data.total_count}}</span></div>
+		<div class="content-heading-info-elem"> <span class="content-heading-info-text">Общий вес: </span><span class="content-heading-info-value">{{data.total_weight}} кг.</span></div>
+		<div class="content-heading-info-elem"> <span class="content-heading-info-text">Общий объем: </span><span class="content-heading-info-value">{{data.total_valume}} м.куб.</span></div>
 	</div>
 </div>
 	
 </template>
+
+<script>
+export default {
+	props: {
+		data: {
+			type: Object,
+		}
+	},
+	setup() {
+		
+	}
+}
+</script>